@@ -4,10 +4,10 @@
   "license": "GPL-3.0",
   "scripts": {
     "ng": "ng",
-    "start": "export NODE_OPTIONS=--dns-result-order=ipv4first && yarn genconfig && yarn run gulp setSecretEnvVars && yarn build:prod && yarn run http-server ./dist -p 4200 --proxy http://localhost:4200?",
+    "start": "export NODE_OPTIONS=--dns-result-order=ipv4first && yarn genconfig && yarn run gulp setSecretEnvVars && yarn build:prod && yarn run http-server ./dist/browser -p 4200 --proxy http://localhost:4200?",
     "start:dev": "export NODE_OPTIONS=--dns-result-order=ipv4first && yarn genconfig && yarn run gulp setSecretEnvVars && ng serve --host 0.0.0.0",
-    "build": "yarn clean && yarn genconfig && ng build",
-    "build:prod": "yarn clean && ng build --configuration production --build-optimizer --aot",
+    "build": "yarn clean && yarn genconfig && ng build --optimization=false --output-hashing=none",
+    "build:prod": "yarn clean && ng build --configuration production",
     "clean": "rm -rf ./dist",
     "test": "yarn genconfig && ng test",
     "coverage": "yarn genconfig && ng test --no-watch --browsers ChromeHeadless --code-coverage",
@@ -21,22 +21,22 @@
     "node": ">= 20"
   },
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/localize": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/platform-server": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/animations": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/localize": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/platform-server": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@babel/runtime": "^7.26.0",
     "@microsoft/applicationinsights-web": "^3.3.3",
     "dotenv": "^16.0.3",
     "fflate": "^0.8.2",
     "lz-string": "^1.5.0",
-    "ngx-cookie-service": "^17.0.0",
+    "ngx-cookie-service": "^18.0.0",
     "polly-js": "^1.8.2",
     "rxjs": "^7.8.1",
     "sass": "^1.83.0",
@@ -45,16 +45,16 @@
     "zone.js": "~0.14"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.11",
-    "@angular-eslint/builder": "17.0.0",
-    "@angular-eslint/eslint-plugin": "17.0.0",
-    "@angular-eslint/eslint-plugin-template": "17.0.0",
-    "@angular-eslint/schematics": "17.0.0",
-    "@angular-eslint/template-parser": "17.0.0",
-    "@angular-eslint/utils": "^17",
-    "@angular/cli": "^17.3.11",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.12",
+    "@angular-eslint/builder": "18.0.0",
+    "@angular-eslint/eslint-plugin": "18.0.0",
+    "@angular-eslint/eslint-plugin-template": "18.0.0",
+    "@angular-eslint/schematics": "18.0.0",
+    "@angular-eslint/template-parser": "18.0.0",
+    "@angular-eslint/utils": "^18.0.0",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/language-service": "^18.2.13",
     "@types/jasmine": "^5.1.5",
     "@types/node": "20",
     "@typescript-eslint/eslint-plugin": "8.18.2",
@@ -78,6 +78,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "node-sass-utils": "^1.1.2",
     "replace-in-file": "^8.3.0",
-    "typescript": "~5.2.0"
+    "typescript": "~5.5.4"
   }
 }
